Rename utils/log.ts → utils/common-log.ts, extract codegen logging to utils/log.ts
Move generic logger engine to common-log.ts and codegen-specific logging
(CodegenTag, CodegenLog, CodegenLogManager, mkCodegenLogger) from
utils/types.ts into utils/log.ts. utils/types.ts now only contains
generic utility types.
