+++ src/api/index.ts
@@ -8,9 +8,8 @@
  */
 
 export type { IrConf, LogicalPromotionConf, TreeShakeConf } from "../typeschema/ir/types";
-export type { LogLevel } from "../utils/log";
-export type { CodegenLog, CodegenLogManager, CodegenTag } from "../utils/types";
-export { mkCodegenLogger } from "../utils/types";
+export type { CodegenLog, CodegenLogManager, CodegenTag, LogLevel } from "../utils/log";
+export { mkCodegenLogger } from "../utils/log";
 export type { APIBuilderOptions, LocalStructureDefinitionConfig } from "./builder";
 export { APIBuilder, prettyReport } from "./builder";
 export type { CSharpGeneratorOptions } from "./writer-generator/csharp/csharp";
