+++ src/typeschema/core/profile-extensions.ts
@@ -7,7 +7,7 @@
 
 import type { FHIRSchemaElement } from "@atomic-ehr/fhirschema";
 import type { Register } from "@root/typeschema/register";
-import type { CodegenLog } from "@root/utils/types";
+import type { CodegenLog } from "@root/utils/log";
 import {
     type CanonicalUrl,
     concatIdentifiers,
