+++ src/utils/types.ts
@@ -1,21 +1,5 @@
-import { type Log, type LogManager, type LoggerOptions, mkLogger } from "./log";
-
 export type CapitalizeFirst<S extends string> = S extends `${infer F}${infer R}` ? `${Uppercase<F>}${R}` : S;
 
 export type IsPrefixed<T extends string> = `is${CapitalizeFirst<T>}`;
 
 export type PartialBy<T, K extends keyof T> = Omit<T, K> & Partial<Pick<T, K>>;
-
-export type CodegenTag =
-    | "#binding"
-    | "#largeValueSet"
-    | "#fieldTypeNotFound"
-    | "#skipCanonical"
-    | "#duplicateSchema"
-    | "#duplicateCanonical"
-    | "#resolveBase";
-
-export type CodegenLog = Log<CodegenTag>;
-export type CodegenLogManager = LogManager<CodegenTag>;
-
-export const mkCodegenLogger = <T extends string = CodegenTag>(opts: LoggerOptions<T> = {}) => mkLogger<T>(opts);
