+++ test/unit/typeschema/utils.ts
@@ -2,8 +2,8 @@ import type { FHIRSchema } from "@atomic-ehr/fhirschema";
 import type { ValueSet } from "@root/fhir-types/hl7-fhir-r4-core";
 import { generateTypeSchemas } from "@root/typeschema";
 import { mkTypeSchemaIndex } from "@root/typeschema/utils";
+import type { CodegenLog, CodegenTag } from "@root/utils/log";
 import { mkLogger } from "@root/utils/log";
-import type { CodegenLog, CodegenTag } from "@root/utils/types";
 import { transformFhirSchema, transformValueSet } from "@typeschema/core/transformer";
 import { type Register, registerFromPackageMetas } from "@typeschema/register";
 import { type CanonicalUrl, enrichFHIRSchema, enrichValueSet, type PackageMeta } from "@typeschema/types";
